Make dicom manager/handlers more flexible
Now, the IncomingDicomManager takes a timeout, a
dicom-to-string function, and a factory function
that should, given an example dicom and a manager,
return a BaseDicomSeriesHandler (or subclass)
that actually handles the dicoms.

The BaseDicomSeriesHandler takes a timeout, a
name, and a manager. No more assumptions about
output directories or anything. The factory
function will also be the one to specify timeout.

The handler has also been modified to allow easy
subclassing to add actual behavior to handling
files and finishing -- just override _handle() and
_finish().

mri.BaseDicom no longer provides series_unique_key
or exam_unique_key.

The leakdown tester has been changed to work with
the new setup and demo subclassing the handler.